requirements.txt
PyYAML
blessings
requests
